Pull TPCCBenchmark functionality up into BenchmarkModule
